requirements.txt
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.29.0
scikit-learn>=1.4.0
fairlearn>=0.10.0
pandas>=2.2.0
numpy>=1.26.0
google-generativeai>=0.5.0
plotly>=5.20.0
fpdf2>=2.7.9
python-multipart>=0.0.9
httpx>=0.27.0
